include module name in function checksum
