+++ qsum/tests/helpers.py
@@ -1,7 +1,7 @@
 # pylint: disable=too-few-public-methods
 import pytest
 
-from qsum import Checksum
+from qsum import Checksum, checksum
 
 STR_CHECKSUM_OBJS = [Checksum('123'),
                      '0001a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3',
@@ -26,3 +26,11 @@ class CustomDict1(dict):
 
 class CustomDict2(dict):
     """Another Custom Class that inherits from dict"""
+
+
+def foo():
+    pass
+
+
+def foo_checksum():
+    return checksum(foo)
